[Refactor] 초대장 뷰 class로 분리
